requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-braces==1.13.0
django-cors-headers==3.1.1
django-extra-fields==2.0.1
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
kombu==4.6.4
more-itertools==7.2.0
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.3
public==2019.4.13
PyJWT==1.7.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
rest-social-auth==2.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.5
values==2019.4.13
vine==1.3.0
zipp==0.6.0
